refactor Util object-> RandomStringGenerator simplify the implementation of the generateRandomString method
