+++ src/main/scala/it/unibo/authsim/library/user/builder/UserBuilder.scala
@@ -3,7 +3,7 @@ package it.unibo.authsim.library.user.builder
 import  it.unibo.authsim.library.user.model.User
 import it.unibo.authsim.library.dsl.policy.checkers.StringPolicyChecker
 import it.unibo.authsim.library.dsl.policy.model.StringPolicies.{CredentialPolicy, PasswordPolicy, UserIDPolicy}
-import it.unibo.authsim.library.user.builder.util.Util.generateRandomString
+import it.unibo.authsim.library.user.builder.util.RandomStringGenerator.generateRandomString
 //ANche qui non sono troppo sicura sull'iso
 abstract class UserBuilder[U]:
   protected var _credentialPolicies: Seq[CredentialPolicy] = Seq.empty
